Refactor note schemas to rename request and response classes for clarity
