+++ src/schemas/__init__.py
@@ -1,10 +1,8 @@
-from schemas.note_versions import (
-    NoteVersionDetailResponseSchema,
-    NoteVersionListResponseSchema,
-)
 from schemas.notes import (
     NoteDetailResponseSchema,
     NoteListResponseSchema,
-    NoteRequestCreateSchema,
-    NoteRequestUpdateSchema,
+    NoteCreateRequestSchema,
+    NoteUpdateRequestSchema,
+    NoteVersionDetailResponseSchema,
+    NoteVersionListResponseSchema,
 )
